refactor: move composer to util package
