+++ src/main/java/com/onixbyte/calendar/component/property/Attachment.java
@@ -25,6 +25,7 @@ package com.onixbyte.calendar.component.property;
 import com.onixbyte.calendar.parameter.FormatType;
 import com.onixbyte.calendar.parameter.InlineEncoding;
 import com.onixbyte.calendar.parameter.ValueDataType;
+import com.onixbyte.calendar.util.PropertyComposer;
 
 import java.net.URI;
 import java.util.Base64;
