+++ src/main/java/com/onixbyte/calendar/component/property/Attendee.java
@@ -23,6 +23,7 @@
 package com.onixbyte.calendar.component.property;
 
 import com.onixbyte.calendar.parameter.*;
+import com.onixbyte.calendar.util.PropertyComposer;
 
 import java.net.URI;
