+++ src/main/java/com/onixbyte/calendar/component/property/ExceptionDateTimes.java
@@ -24,6 +24,7 @@ package com.onixbyte.calendar.component.property;
 
 import com.onixbyte.calendar.parameter.TimeZoneIdentifier;
 import com.onixbyte.calendar.parameter.ValueDataType;
+import com.onixbyte.calendar.util.PropertyComposer;
 
 import java.time.ZonedDateTime;
