+++ src/main/java/com/onixbyte/calendar/component/property/Location.java
@@ -24,6 +24,7 @@ package com.onixbyte.calendar.component.property;
 
 import com.onixbyte.calendar.parameter.AlternateTextRepresentation;
 import com.onixbyte.calendar.parameter.Language;
+import com.onixbyte.calendar.util.PropertyComposer;
 
 /**
  * Represents the iCalendar {@code LOCATION} property, which defines the intended venue for a
